% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rotated.R
\name{as.rotated}
\alias{as.rotated}
<<<<<<< HEAD
\title{as.rotated}
=======
\title{Rotate a band matrix to get the rotated row-wised matrix associated.}
>>>>>>> 2b739f4834cb5b76dc34ce5b9deb93e350d03fcd
\usage{
as.rotated(M, l = 0, u = 0)
}
\arguments{
\item{M}{Band square matrix.}

\item{l}{lower bandwith.}

\item{u}{upper bandwith.}
}
\value{
Rotated matrix.
}
<<<<<<< HEAD
\description{
Rotate a band matrix to get the rotated row-wised matrix associated.
}
=======
>>>>>>> 2b739f4834cb5b76dc34ce5b9deb93e350d03fcd
\examples{
n=2000;
A=diag(4)
A[2,3]=2
ref=as.rotated(A,l=0,u=1)

}

