% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bandsolve4}
\alias{bandsolve4}
\title{Fast solver for linears systems involving symmetric band matrix.}
\usage{
bandsolve4(D0data, D1data, D2data, D3data, D4data, bdata)
}
\arguments{
\item{D0data}{Diagonal as a vector of length n.}

\item{bdata}{Right hand side of the linear system}

\item{D1Data}{First subdiagonal and superdiagonal as a vector of length n-1.}

\item{D2Data}{Second subdiagonal and superdiagonal as a vector of length n-2.}

\item{D3Data}{Third subdiagonal and superdiagonal as a vector of length n-3.}

\item{D4Data}{Fourth subdiagonal and superdiagonal as a vector of length n-4.}
}
\value{
Solution of the linear system.
}
\description{
Fast solver for linears systems involving symmetric band matrix.
}
\examples{
n=2000;
D0=runif(n);
D1=-0.2*runif(n-1);
D2=0.4*runif(n-2);
D3=-0.1*runif(n-3);
D4=0.3*runif(n-4);
b=runif(n)
ref=bandsolve4(D0,D1,D2,D3,D4,b)
}

