% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bandsolveK}
\alias{bandsolveK}
\title{Fast solver for linears systems involving symmetric band matrix of length k.}
\usage{
bandsolveK(Ddata, bdata)
}
\arguments{
\item{Ddata}{Rotated row-wised matrix of dimensions n*k}

\item{bdata}{Right hand side of the linear system}
}
\value{
Solution of the linear system.
}
\description{
Fast solver for linears systems involving symmetric band matrix of length k.
}
\examples{
n=2000;
D0=runif(n);
D1=-0.2*runif(n-1);
D=cbind(D0,c(D1,0))
b=runif(n)
ref=bandsolveK(D,b)
}

