% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LDU}
\alias{LDU}
\title{LDU}
\arguments{
\item{D}{Rotated row-wised matrix of dimensions n*(l+1+u), where the l first column corresponding to the sub-diagonal elements, the column l+1 corresponding to the diagonal and the u next to the super-diagonal elements.}

\item{l}{Lower bandwidth.}

\item{u}{Upper bandwidth.}
}
\value{
List with D as solution of our LDU decomposition.
}
\description{
Fast inplace LDU decomposition of band matrix.
}
\examples{
n=2000;
D0=0.3*runif(n-1);
D1=3+runif(n);
D2=-0.2*runif(n-1);
D3=runif(n-2)
D=cbind(c(D0,0),D1,c(D2,0),c(D3,0,0))
res=LDU(D,l=1,u=2)
}

