% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.mat.R
\name{rot.mat}
\alias{rot.mat}
\title{rot.mat}
\usage{
rot.mat(R)
}
\arguments{
\item{R}{Rotated matrix.}
}
\value{
Band square matrix.
}
\description{
Get back a symmetric square matrix based on his rotated row-wised version. 
The rotated form of the input is such each column correspond to a diagonal, where the first column is the main diagonal and next ones are the upper/lower-diagonal.
To match dimension, last element of these columns are discarded.
}
\examples{
require(bandsolve)
D0=1:5;
D1=c(0,1,0,0);
D2=rep(2,3);

ref=rot.mat(cbind(D0,c(D1,0),c(D2,0,0)))
ref
mat.rot(rot.mat(cbind(D0,c(D1,0),c(D2,0,0))))

}

