% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandsolve.R
\docType{package}
\name{bandsolve-package}
\alias{bandsolve}
\alias{bandsolve-package}
\title{Generate R documentation from inline comments.}
\description{
Roxygen2 allows you to write documentation in comment blocks co-located
with code.
}

