% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandsolve.R
\docType{package}
\name{bandsolve-package}
\title{\code{\link{bandsolve}},\code{\link{LDU}},\code{\link{LDL}},\code{\link{rot.as.mat}},\code{\link{as.rotated}}}
\description{
\code{\link{bandsolve}},\code{\link{LDU}},\code{\link{LDL}},\code{\link{rot.as.mat}},\code{\link{as.rotated}}
}

