% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bandsolve}
\alias{bandsolve}
\title{bandsolve}
\arguments{
\item{D}{Rotated row-wised matrix.}

\item{b}{Right hand side of the linear system}
}
\value{
List with components with the solution of the linear system and the rotated matrixes L and U.
}
\description{
Fast solver for linears systems involving band matrix.
}
\examples{
n=2000;
D0=3+runif(n);
D1=-0.2*runif(n-1);
D=cbind(D0,c(D1,0))
b=runif(n)
sol=bandsolve(D,b,l=1,u=0)
}

