% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bandsolve.R
\name{bandsolve}
\alias{bandsolve}
\title{Fast solver for linears systems involving symmetric band matrix.}
\usage{
bandsolve(D, b)
}
\arguments{
\item{D}{Matrice rotationné}

\item{b}{Right hand side of the linear system}
}
\value{
Solution of the linear system.
}
\description{
Fast solver for linears systems involving symmetric band matrix.
}
\examples{
n=2000;
D0=runif(n);
D1=-0.2*runif(n-1);
D2=0.3*runif(n-2);
D3=0.4*runif(n-3);
D4=-0.1*runif(n-4);
b=runif(n)
ref=bandsolve(D0,D1,D2,D3,D4,b)
}

